Implement common layout for job post modal tabs
